+++ Menu/src/main/java/db/Product.java
@@ -23,7 +23,7 @@ public class Product extends AbstractProduct implements java.io.Serializable {
 
 	/** full constructor */
 	public Product(Integer id, Category category, String name,
-			String descryption, Integer raiting, Double price, String imagePath, Integer minut, Integer calories) {
+			String descryption, Double raiting, Integer price, String imagePath, Integer minut, Integer calories) {
 		super(id, category, name, descryption, raiting, price, imagePath, minut, calories) ;
 	}
